some more adaptive sensing

+++ TaskyApp/app/src/main/java/si/uni_lj/fri/taskyapp/sensor/Constants.java
@@ -4,8 +4,8 @@ package si.uni_lj.fri.taskyapp.sensor;
  * Created by urgas9 on 10. 01. 2016.
  */
 public class Constants {
-    public static final int APPROXIMATE_INTERVAL_MILLIS = 60000 * 30;//30000;
-    public static final int MIN_INTERVAL_MILLIS = 1000 * 30 * 5; // Five mins is the minimum time required between two sensing
+    public static final long APPROXIMATE_INTERVAL_MILLIS = 60000 * 30;//30000;
+    public static final long MIN_INTERVAL_MILLIS = 1000 * 30 * 5; // Five mins is the minimum time required between two sensing
     public static final long MAX_INTERVAL_WITHOUT_SENSING_DATA_IN_MILLIS = (long)(APPROXIMATE_INTERVAL_MILLIS * 1.2);
     public static final long SENSING_WINDOW_LENGTH_MILLIS = 10 * 1000;
